checking for shipment in packing slip before deleting. if shipment exists, deletion is not allowed

+++ server/src/packingSlip/controller.js
@@ -142,6 +142,8 @@ async function deletePackingSlip(req, res) {
   handler(
     async () => {
       const { pid } = req.params;
+      const doc = PackingSlip.findOne({ _id: pid }).lean();
+      if (doc.shipment) return [{ status: 400, message: 'That packing slip has already been shipped.' }];
 
       await PackingSlip.deleteOne({ _id: pid });
       return [null];
